requirements.txt
miflora >= 0.7.1
bluepy >= 1.3.0
requests >= 2.26.0
python-dotenv >= 0.19.0
